docs: rename PrivacyFilter to RedactableValues
