+++ src/templates/examples/TemplateB/index.tsx
@@ -1,9 +1,9 @@
-import { TemplateBWithPrivacyFilter } from "./TemplateBWithPrivacyFilter";
+import { TemplateBWithRedactableValues } from "./TemplateBWithRedactableValues";
 
 export const TemplateBTemplates = [
   {
     id: "template-b",
-    label: "Template B with privacy filter",
-    template: TemplateBWithPrivacyFilter,
+    label: "Template B with RedactableValues",
+    template: TemplateBWithRedactableValues,
   },
 ];
